refactor(zero-permissions): extract permissions into its own package
`zero-schema` had a production dependency on `zql` via permissions.ts
importing `ExpressionBuilder`, `Query`, and `newExpressionBuilder`. This
created a cycle since `zql` tests import `zero-schema` builders.

Move `definePermissions` and related exports to a new `zero-permissions`
package that explicitly depends on both `zero-schema` and `zql`,
breaking the cycle. Update all consumers in `zero-client`, `zero-cache`,
and `zql` to import from the new package.
