+++ packages/zero-cache/src/scripts/schema-permissions-test.ts
@@ -1,10 +1,10 @@
-import {relationships} from '../../../zero-schema/src/builder/relationship-builder.ts';
-import {createSchema} from '../../../zero-schema/src/builder/schema-builder.ts';
-import {string, table} from '../../../zero-schema/src/builder/table-builder.ts';
 import {
   ANYONE_CAN_DO_ANYTHING,
   definePermissions,
-} from '../../../zero-schema/src/permissions.ts';
+} from '../../../zero-permissions/src/permissions.ts';
+import {relationships} from '../../../zero-schema/src/builder/relationship-builder.ts';
+import {createSchema} from '../../../zero-schema/src/builder/schema-builder.ts';
+import {string, table} from '../../../zero-schema/src/builder/table-builder.ts';
 
 const member = table('member')
   .columns({
