+++ packages/zero-cache/src/services/mutagen/mutagen.authz.pg.test.ts
@@ -2,6 +2,11 @@ import {beforeEach, describe, expect} from 'vitest';
 import {testLogConfig} from '../../../../otel/src/test-log-config.ts';
 import {h128} from '../../../../shared/src/hash.ts';
 import {createSilentLogContext} from '../../../../shared/src/logging-test-utils.ts';
+import {
+  ANYONE_CAN,
+  definePermissions,
+  NOBODY_CAN,
+} from '../../../../zero-permissions/src/permissions.ts';
 import * as MutationType from '../../../../zero-protocol/src/mutation-type-enum.ts';
 import {createSchema} from '../../../../zero-schema/src/builder/schema-builder.ts';
 import {
@@ -11,11 +16,6 @@ import {
   string,
   table,
 } from '../../../../zero-schema/src/builder/table-builder.ts';
-import {
-  ANYONE_CAN,
-  definePermissions,
-  NOBODY_CAN,
-} from '../../../../zero-schema/src/permissions.ts';
 import type {Schema as ZeroSchema} from '../../../../zero-types/src/schema.ts';
 import type {ExpressionBuilder} from '../../../../zql/src/query/expression.ts';
 import type {Row} from '../../../../zql/src/query/query.ts';
