+++ packages/zero-cache/src/services/view-syncer/view-syncer-test-util.ts
@@ -4,6 +4,10 @@ import {testLogConfig} from '../../../../otel/src/test-log-config.ts';
 import {h128} from '../../../../shared/src/hash.ts';
 import {createSilentLogContext} from '../../../../shared/src/logging-test-utils.ts';
 import {Queue} from '../../../../shared/src/queue.ts';
+import {
+  ANYONE_CAN_DO_ANYTHING,
+  definePermissions,
+} from '../../../../zero-permissions/src/permissions.ts';
 import type {AST} from '../../../../zero-protocol/src/ast.ts';
 import {type ClientSchema} from '../../../../zero-protocol/src/client-schema.ts';
 import type {
@@ -26,10 +30,6 @@ import {
   table,
 } from '../../../../zero-schema/src/builder/table-builder.ts';
 import type {PermissionsConfig} from '../../../../zero-schema/src/compiled-permissions.ts';
-import {
-  ANYONE_CAN_DO_ANYTHING,
-  definePermissions,
-} from '../../../../zero-schema/src/permissions.ts';
 import type {ExpressionBuilder} from '../../../../zql/src/query/expression.ts';
 import {
   CREATE_STORAGE_TABLE,
