+++ packages/zero-permissions/src/env.d.ts
@@ -1,0 +1,3 @@
+interface ImportMeta {
+  readonly env?: {readonly VITEST?: string | boolean};
+}
