+++ packages/zero-permissions/src/permissions.test.ts
@@ -1,8 +1,8 @@
 import {expect, test} from 'vitest';
+import {createSchema} from '../../zero-schema/src/builder/schema-builder.ts';
+import {column, table} from '../../zero-schema/src/builder/table-builder.ts';
 import type {Schema as ZeroSchema} from '../../zero-types/src/schema.ts';
 import type {ExpressionBuilder} from '../../zql/src/query/expression.ts';
-import {createSchema} from './builder/schema-builder.ts';
-import {column, table} from './builder/table-builder.ts';
 import {definePermissions} from './permissions.ts';
 
 const {string} = column;
