--- packages/zero-schema/src/mod.ts
@@ -1,3 +1,0 @@
-export {createSchema} from './builder/schema-builder.ts';
-export {definePermissions} from './permissions.ts';
-export {type TableSchema} from './table-schema.ts';
